+++ gems/aws-sdk-ec2/CHANGELOG.md
@@ -1,6 +1,11 @@
 Unreleased Changes
 ------------------
 
+1.644.0 (2026-09-04)
+------------------
+
+* Feature - Adds support for ValidateSecurityGroupQuotasForInterface, an API that specifically authorized AWS services use to validate security group rule quotas before creating an elastic network interface.
+
 1.643.0 (2026-09-02)
 ------------------